fix(preload): collapse index.d.ts into type-checked api-types.ts
Renames src/preload/api-types.d.ts to .ts and deletes the duplicate
src/preload/index.d.ts so PreloadApi becomes the single type-checked
source of truth for window.api. Previously, skipLibCheck silently
widened unresolved names in the .d.ts to any, which is how #1186 shipped
a broken IPC signature past typecheck. Adds a CI guard blocking new
project-owned .d.ts under src/preload/ or src/shared/, documents the
.ts-over-.d.ts convention in AGENTS.md and CONTRIBUTING.md, and retargets
tsconfig.web/tc.web includes to the renamed file. Full rationale in
docs/preload-typecheck-hole.md.
